+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 Notable changes will be documented here
 
+## 0.0.22
+
+- link-checker-service:
+  - a new optional memory-limited cache based on [github.com/dgraph-io/ristretto](https://github.com/dgraph-io/ristretto)
+    - 
+    - run the service with `--cacheUseRistretto`
+    - see the options in [.link-checker-service.toml](.link-checker-service.toml)
+
 ## 0.9.21
 
 - link-checker-service: